Refactor `ConcurrentHashMap` for improved concurrency handling.
Added comments to clarify lock usage and ensured thread-safe size management. Updated `Size()` return type to `int64` for consistency and scalability. Expanded test coverage with new cases for concurrent operations and edge scenarios.

+++ maps/map.go
@@ -4,5 +4,5 @@ type Map[K comparable, V any] interface {
 	Put(key K, value V)   // Inserts or updates a key-value pair
 	Get(key K) (V, error) // Retrieves the value associated with a key
 	Remove(key K) error   // Removes a key-value pair
-	Size() int            // Returns the number of key-value pairs
+	Size() int64          // Returns the number of key-value pairs
 }
